bebeğim/search_backend/requirements.txt
openai~=0.28.0
tiktoken~=0.7.0
python-dotenv~=1.0.0
azure-storage-blob
azure-search-documents
fastapi
uvicorn
numpy
openai[datalib]
pydantic


